// client/app/models/system-config.ts
export class SystemConfig {
    systemConfigId: string;
    systemConfigName: string;
    systemConfigFieldName: string;
    systemConfigType: string;
    systemConfigDescription: string;
    systemConfigInputType: string;
    systemConfigValue: string;
    systemConfigOptionValues: string;
    systemConfigControlType: string;
    systemConfigLabelWidth: number;
    systemConfigInputWidth: number;
    systemConfigRequired: boolean;
    systemConfigInline: boolean;
    systemConfigOrder: number;
    isActive: boolean;
    createdBy: any;
    updatedBy: any;
    createdAt: Date;
    updatedAt: Date;
}